+++ R/httr2-package.R
@@ -15,3 +15,4 @@ the$cache_throttle <- list()
 the$token_cache <- new_environment()
 the$last_response <- NULL
 the$last_request <- NULL
+the$pool_pollers <- new_environment()